requirements.txt
Django>=4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
nltk>=3.8
scikit-learn>=1.3.0
numpy>=1.24.0
spacy>=3.7.0
requests>=2.31.0
reportlab>=3.7.0
django-allauth>=0.63.0
pymysql>=1.1.0
gunicorn>=23.0.0
whitenoise>=6.8.2
dj-database-url>=2.2.0
